requirements.txt
kafka-python==2.2.15
pyspark==4.0.0
pandas==2.3.0
numpy==2.3.1